GameServer.Accounts (game_server v1.0.231)

The Accounts context.

Functions

attach_device_to_user(user, device_id)

Attach a device_id to an existing user record. Returns {:ok, user} or {:error, changeset} if the device_id is already used.

broadcast_user_update(user)

Broadcast that the given user has been updated.

This helper is intentionally small and only broadcasts a compact payload intended for client consumption through the user_updates:<id> topic.

change_user_display_name(user, attrs \\ %{})

Returns an %Ecto.Changeset{} for changing the user display_name.

change_user_email(user, attrs \\ %{}, opts \\ [])

Returns an %Ecto.Changeset{} for changing the user email.

See GameServer.Accounts.User.email_changeset/3 for a list of supported options.

Examples

iex> change_user_email(user)
%Ecto.Changeset{data: %User{}}

change_user_password(user, attrs \\ %{}, opts \\ [])

Returns an %Ecto.Changeset{} for changing the user password.

See GameServer.Accounts.User.password_changeset/3 for a list of supported options.

Examples

iex> change_user_password(user)
%Ecto.Changeset{data: %User{}}

change_user_registration(user, attrs \\ %{})

confirm_user(user)

Confirms a user's email by setting confirmed_at timestamp.

Examples

iex> confirm_user(user)
{:ok, %User{}}

count_search_users(query)

Count users matching a text query (email or display_name). Returns integer.

delete_user(user)

Deletes a user and associated resources.

Returns {:ok, user} on success or {:error, changeset} on failure.

delete_user_session_token(token)

Deletes the signed token with the given context.

deliver_login_instructions(user, magic_link_url_fun)

Delivers the magic link login instructions to the given user.

deliver_user_confirmation_instructions(user, confirmation_url_fun)

deliver_user_update_email_instructions(user, current_email, update_email_url_fun)

Delivers the update email instructions to the given user.

Examples

iex> deliver_user_update_email_instructions(user, current_email, &url(~p"/users/settings/confirm-email/#{&1}"))
{:ok, %{to: ..., body: ...}}

detach_device_from_user(user)

Clears a device_id from the user if present. Returns {:ok, user} or {:error, changeset} if update fails.

device_auth_enabled?()

Returns true when device-based auth is enabled. This checks the application config :game_server, :device_auth_enabled and falls back to the environment variable DEVICE_AUTH_ENABLED. If neither is set, device auth is enabled by default.

find_or_create_from_apple(attrs)

Finds a user by Apple ID or creates a new user from OAuth data.

Examples

iex> find_or_create_from_apple(%{apple_id: "123", email: "user@example.com"})
{:ok, %User{}}

find_or_create_from_device(device_id, attrs \\ %{})

Finds or creates a user associated with the given device_id.

If a user already exists with the device_id we return it. Otherwise we create an anonymous confirmed user and attach the device_id.

find_or_create_from_discord(attrs)

Finds a user by Discord ID or creates a new user from OAuth data.

Examples

iex> find_or_create_from_discord(%{discord_id: "123", email: "user@example.com"})
{:ok, %User{}}

find_or_create_from_facebook(attrs)

Finds a user by Facebook ID or creates a new user from OAuth data.

Examples

iex> find_or_create_from_facebook(%{facebook_id: "123", email: "user@example.com"})
{:ok, %User{}}

find_or_create_from_google(attrs)

Finds a user by Google ID or creates a new user from OAuth data.

Examples

iex> find_or_create_from_google(%{google_id: "123", email: "user@example.com"})
{:ok, %User{}}

find_or_create_from_steam(attrs)

Finds a user by Steam ID or creates a new user from Steam OpenID data.

Examples

iex> find_or_create_from_steam(%{steam_id: "12345", email: "user@example.com"})
{:ok, %User{}}

generate_user_session_token(user)

Generates a session token.

get_user!(id)

Gets a single user.

Raises Ecto.NoResultsError if the User does not exist.

Examples

iex> get_user!(123)
%User{}

iex> get_user!(456)
** (Ecto.NoResultsError)

get_user_by_apple_id(apple_id)

Get a user by their Apple ID.

Returns %User{} or nil.

get_user_by_discord_id(discord_id)

Get a user by their Discord ID.

Returns %User{} or nil.

get_user_by_email(email)

Gets a user by email.

Examples

iex> get_user_by_email("foo@example.com")
%User{}

iex> get_user_by_email("unknown@example.com")
nil

get_user_by_email_and_password(email, password)

Gets a user by email and password.

Examples

iex> get_user_by_email_and_password("foo@example.com", "correct_password")
%User{}

iex> get_user_by_email_and_password("foo@example.com", "invalid_password")
nil

get_user_by_facebook_id(facebook_id)

Get a user by their Facebook ID.

Returns %User{} or nil.

get_user_by_google_id(google_id)

Get a user by their Google ID.

Returns %User{} or nil.

get_user_by_magic_link_token(token)

Gets the user with the given magic link token.

get_user_by_session_token(token)

Gets the user with the given signed token.

If the token is valid {user, token_inserted_at} is returned, otherwise nil is returned.

get_user_by_steam_id(steam_id)

Get a user by their Steam ID (steam_id).

Returns %User{} or nil.

link_account(user, attrs, provider_id_field, changeset_fn)

Link an OAuth provider to an existing user account. Updates the user via the provider's oauth changeset while being careful not to overwrite existing email or avatars.

Example: link_account(user, %{discord_id: "123", profile_url: "https://..."}, :discord_id, &User.discord_oauth_changeset/2)

login_user_by_magic_link(token)

Logs the user in by magic link.

There are three cases to consider:

The user has already confirmed their email. They are logged in and the magic link is expired.

The user has not confirmed their email and no password is set. In this case, the user gets confirmed, logged in, and all tokens - including session ones - are expired. In theory, no other tokens exist but we delete all of them for best security practices.

The user has not confirmed their email but a password is set. This cannot happen in the default implementation but may be the source of security pitfalls. See the "Mixing magic link and password registration" section of mix help phx.gen.auth.

register_user(attrs)

Registers a user.

Examples

iex> register_user(%{field: value})
{:ok, %User{}}

iex> register_user(%{field: bad_value})
{:error, %Ecto.Changeset{}}

search_users(query, opts \\ [])

Search users by email or display name (case-insensitive, partial match).

Returns a list of User structs (up to 50 results).

sudo_mode?(user, minutes \\ -20)

Checks whether the user is in sudo mode.

The user is in sudo mode when the last authentication was done no further than 20 minutes ago. The limit can be given as second argument in minutes.

unlink_provider(user, provider)

Unlink an OAuth provider from a user's account.

provider should be one of :discord, :apple, :google, :facebook. This will return {:ok, user} when successful or {:error, reason}.

Guard: we only allow unlinking when the user will still have at least one other social provider remaining. This prevents users losing all social logins unexpectedly.

update_user_admin(user, attrs)

Admin-level user update wrapper.

This function applies the User.admin_changeset/2 then updates the user and broadcasts the update on success. It returns the same tuple shape as Repo.update/1 so callers can pattern-match as before.

update_user_display_name(user, attrs)

Updates the user's display name and broadcasts the change.

update_user_email(user, token)

Updates the user email using the given token.

If the token matches, the user email is updated and the token is deleted.

update_user_password(user, attrs)

Updates the user password.

Returns a tuple with the updated user, as well as a list of expired tokens.

Examples

iex> update_user_password(user, %{password: ...})
{:ok, {%User{}, [...]}}

iex> update_user_password(user, %{password: "too short"})
{:error, %Ecto.Changeset{}}